Moved page to main page. Final product i think.
More additions I would make:
- Sound effects for attacks
- Victory sounds
- a hard mode (adjust the damage)
- an endurance mode (adjust the healing)
- a math mode (solve math problems w/ turn timer)
- mac and cheese visual counter
