+++ battle_showdown_challenge/src/app/globals.css
@@ -17,17 +17,19 @@
 }
 
 @theme {
-  --animate-slidein: slidein 1s ease-in-out 1;
-  --animate-slideinl: slideinl 1s ease-in-out 1;
-  --animate-slideinu: slideinu 1s ease-in-out 1;
-  --animate-fork: fork 1.2s linear 1;
+  --animate-slidein: slidein 1s ease-in-out 1 forwards;
+  --animate-slideinl: slideinl 1s ease-in-out 1 forwards;
+  --animate-slideinu: slideinu 1s ease-in-out 1 forwards;
+  --animate-fork: fork 1.2s linear 1 forwards;
   --animate-idle: idle 5s ease-in-out infinite;
-  --animate-mac: mac 1.2s linear 1;
-  --animate-snowb: snowball 1.2s linear 1;
-  --animate-fall: fall 1.2s linear 1;
+  --animate-start: start 5s ease-in-out infinite;
+  --animate-mac: mac 1.2s linear 1 forwards;
+  --animate-snowb: snowball 1.2s linear 1 forwards;
+  --animate-fall: fall 1.2s linear 1 forwards;
+  --animate-slideind: slideind 1s ease-in-out 1 forwards;
 
-  --animate-buildm: buildM 1.2s linear 1;
-  --animate-buildp: buildP 1.2s linear 1;
+  --animate-buildm: buildM 1.2s linear 1 forwards;
+  --animate-buildp: buildP 1.2s linear 1 forwards;
   
 
   @keyframes slidein {
@@ -50,7 +52,16 @@
 
   @keyframes slideinu {
     from {
-      transform:translateY(-2000px);
+      transform:translateY(-4000px);
+    }
+    to {
+      transform:translateY(0px);
+    }
+  }
+
+  @keyframes slideind {
+    from {
+      transform:translateY(2000px);
     }
     to {
       transform:translateY(0px);
@@ -123,10 +134,22 @@
       transform:translateY(0px) translateX(0px);
     }
     50% {
-      transform:translateY(-10px) translateX(-10px);
+      transform:translateY(-10px) translateX(10px);
     }
     100% {
       transform:translateY(0px) translateX(0px);
+    }
+  }
+
+  @keyframes start  {
+    0% {
+      transform:translateY(0px);
+    }
+    50% {
+      transform:translateY(-10px);
+    }
+    100% {
+      transform:translateY(0px);
     }
   }
